Clean old files. Refactor.
