--- scripts/deploy/osmosis/devIndex.ts
@@ -1,6 +1,0 @@
-import { taskRunner } from '../base'
-import { osmosisDevnet } from './devnetConfig'
-
-void (async function () {
-  await taskRunner(osmosisDevnet)
-})()
