--- scripts/deploy/osmosis/mainIndex.ts
@@ -1,6 +1,0 @@
-import { taskRunner } from '../base'
-import { osmosisMainnet } from './mainnetConfig'
-
-void (async function () {
-  await taskRunner(osmosisMainnet)
-})()
