+++ scripts/deploy/osmosis/mainet-index.ts
@@ -1,10 +1,10 @@
 import { taskRunner } from '../base'
-import { osmosisMainnet } from './NEW_mainnetConfig'
+import { osmosisMainnetConfig } from './mainnet-config'
 
 void (async function () {
   await taskRunner({
     config: {
-      ...osmosisMainnet,
+      ...osmosisMainnetConfig,
       deployerMnemonic: 'TO BE INSERTED AT TIME OF DEPLOYMENT',
       multisigAddr: 'osmo14w4x949nwcrqgfe53pxs3k7x53p0gvlrq34l5n',
     },
