--- scripts/deploy/osmosis/multisig.ts
@@ -1,6 +1,0 @@
-import { taskRunner } from '../base'
-import { osmosisTestMultisig } from './testnetConfig'
-
-void (async function () {
-  await taskRunner(osmosisTestMultisig)
-})()
