+++ scripts/deploy/osmosis/testnet-config.ts
@@ -1,40 +1,247 @@
-import { DeploymentConfig, VaultType } from '../../types/config.rover'
+import {
+  DeploymentConfig,
+  AssetConfig,
+  OracleConfig,
+  VaultConfig,
+  VaultType,
+} from '../../types/NEW_config'
 
-// Note: since osmo-test-5 upgrade, testnet and mainnet denoms are no longer the same. Reference asset info here: https://docs.osmosis.zone/osmosis-core/asset-info/
+// assets based off of OSMO-TEST-5: https://docs.osmosis.zone/osmosis-core/asset-info/
 const uosmo = 'uosmo'
 const aUSDC = 'ibc/6F34E1BD664C36CE49ACC28E60D62559A5F96C4F9A6CCE4FC5A67B2852E24CFE' // axelar USDC
-// const atom = 'ibc/A8C2D23A1E6F95DA4E48BA349667E322BD7A6C996D8A4AAE8BA72E190F3D1477'
+const atom = 'ibc/A8C2D23A1E6F95DA4E48BA349667E322BD7A6C996D8A4AAE8BA72E190F3D1477'
+const mars = 'ibc/2E7368A14AC9AB7870F32CFEA687551C5064FA861868EDF7437BC877358A81F9'
+const usdcOsmo = 'gamm/pool/5'
+const atomOsmo = 'gamm/pool/12'
 
-const ausdcOsmoPool = 'gamm/pool/5'
-// const atomOsmoPool = 'gamm/pool/12'
+const protocolAdminAddr = 'osmo14w4x949nwcrqgfe53pxs3k7x53p0gvlrq34l5n'
 
-// All vaults below are ONE day vaults
-const atomOsmoVault = 'osmo1m45ap4rq4m2mfjkcqu9ks9mxmyx2hvx0cdca9sjmrg46q7lghzqqhxxup5'
-const ausdcOsmoVault = 'osmo1l3q4mrhkzjyernjhg8lz2t52ddw589y5qc0z7y8y28h6y5wcl46sg9n28j'
-
-const ATOM_OSMO_Config = (addr: string) => ({
-  addr,
-  deposit_cap: { denom: aUSDC, amount: '1000000000' }, // 1000 atom
-  max_loan_to_value: '0.63',
-  liquidation_threshold: '0.65',
-  whitelisted: true,
-})
-const aUSDC_OSMO_Config = (addr: string) => ({
-  addr,
-  deposit_cap: { denom: aUSDC, amount: '1000000000' }, // 1000 atom
-  max_loan_to_value: '0.63',
-  liquidation_threshold: '0.65',
-  whitelisted: true,
-})
+// note the following addresses are all 'mars' bech32 prefix
+const safetyFundAddr = 'mars1s4hgh56can3e33e0zqpnjxh0t5wdf7u3pze575'
+const feeCollectorAddr = 'mars17xpfvakm2amg962yls6f84z3kell8c5ldy6e7x'
 
 const defaultCreditLine = '100000000000'
 
+export const osmoAsset: AssetConfig = {
+  credit_manager: {
+    whitelisted: true,
+  },
+  symbol: 'OSMO',
+  denom: uosmo,
+  liquidation_bonus: {
+    max_lb: '0.05',
+    min_lb: '0',
+    slope: '2',
+    starting_lb: '0',
+  },
+  protocol_liquidation_fee: '0.5',
+  liquidation_threshold: '0.61',
+  max_loan_to_value: '0.59',
+  red_bank: {
+    borrow_enabled: true,
+    deposit_enabled: true,
+  },
+  deposit_cap: '2500000000000',
+  reserve_factor: '0.2',
+  interest_rate_model: {
+    optimal_utilization_rate: '0.8',
+    base: '0',
+    slope_1: '0.2',
+    slope_2: '2',
+  },
+}
+
+export const atomAsset: AssetConfig = {
+  credit_manager: {
+    whitelisted: true,
+  },
+  symbol: 'ATOM',
+  denom: atom,
+  liquidation_bonus: {
+    max_lb: '0.05',
+    min_lb: '0',
+    slope: '2',
+    starting_lb: '0',
+  },
+  protocol_liquidation_fee: '0.5',
+  liquidation_threshold: '0.7',
+  max_loan_to_value: '0.68',
+  red_bank: {
+    borrow_enabled: true,
+    deposit_enabled: true,
+  },
+  deposit_cap: '100000000000',
+  reserve_factor: '0.2',
+  interest_rate_model: {
+    optimal_utilization_rate: '0.8',
+    base: '0',
+    slope_1: '0.2',
+    slope_2: '2',
+  },
+}
+
+export const USDCAsset: AssetConfig = {
+  credit_manager: {
+    whitelisted: true,
+  },
+  symbol: 'aUSDC',
+  denom: aUSDC,
+  liquidation_bonus: {
+    max_lb: '0.05',
+    min_lb: '0',
+    slope: '2',
+    starting_lb: '0',
+  },
+  protocol_liquidation_fee: '0.5',
+  liquidation_threshold: '0.75',
+  max_loan_to_value: '0.74',
+  red_bank: {
+    borrow_enabled: true,
+    deposit_enabled: true,
+  },
+  deposit_cap: '500000000000',
+  reserve_factor: '0.2',
+  interest_rate_model: {
+    optimal_utilization_rate: '0.8',
+    base: '0',
+    slope_1: '0.2',
+    slope_2: '2',
+  },
+}
+
+export const usdcOsmoVault: VaultConfig = {
+  symbol: 'usdcOsmoVault',
+  vault: {
+    addr: 'osmo1l3q4mrhkzjyernjhg8lz2t52ddw589y5qc0z7y8y28h6y5wcl46sg9n28j',
+    deposit_cap: {
+      denom: aUSDC,
+      amount: '1000000000',
+    },
+    liquidation_threshold: '0.65',
+    max_loan_to_value: '0.63',
+    whitelisted: true,
+  },
+}
+
+export const atomOsmoVault: VaultConfig = {
+  symbol: 'atomOsmoVault',
+  vault: {
+    addr: 'osmo1m45ap4rq4m2mfjkcqu9ks9mxmyx2hvx0cdca9sjmrg46q7lghzqqhxxup5',
+    deposit_cap: {
+      denom: aUSDC,
+      amount: '1000000000',
+    },
+    liquidation_threshold: '0.65',
+    max_loan_to_value: '0.63',
+    whitelisted: true,
+  },
+}
+
+export const osmoOracle: OracleConfig = {
+  denom: uosmo,
+  price_source: {
+    fixed: {
+      price: '1',
+    },
+  },
+}
+
+export const atomOracle: OracleConfig = {
+  denom: atom,
+  price_source: {
+    geometric_twap: {
+      downtime_detector: { downtime: 'Duration30m', recovery: 7200 },
+      window_size: 1800,
+      pool_id: 12,
+    },
+  },
+}
+export const USDCOracle: OracleConfig = {
+  denom: aUSDC,
+  price_source: {
+    staked_geometric_twap: {
+      transitive_denom: uosmo,
+      pool_id: 5,
+      window_size: 1800,
+      downtime_detector: { downtime: 'Duration30m', recovery: 7200 },
+    },
+  },
+}
+
+export const usdcOsmoOracle: OracleConfig = {
+  denom: usdcOsmo,
+  price_source: {
+    xyk_liquidity_token: {
+      pool_id: 5,
+    },
+  },
+}
+
+export const atomOsmoOracle: OracleConfig = {
+  denom: atomOsmo,
+  price_source: {
+    xyk_liquidity_token: {
+      pool_id: 12,
+    },
+  },
+}
+
+const testActions = {
+  vault: {
+    depositAmount: '1000000',
+    withdrawAmount: '1000000',
+    mock: {
+      config: {
+        deposit_cap: { denom: aUSDC, amount: '100000000' }, // 100 usdc
+        liquidation_threshold: '0.585',
+        max_loan_to_value: '0.569',
+        whitelisted: true,
+      },
+      vaultTokenDenom: uosmo,
+      type: VaultType.LOCKED,
+      lockup: { time: 900 }, // 15 mins
+      baseToken: usdcOsmo,
+    },
+  },
+  borrowAmount: '10',
+  repayAmount: '11',
+  depositAmount: '100',
+  lendAmount: '10',
+  reclaimAmount: '5',
+  secondaryDenom: aUSDC,
+  startingAmountForTestUser: '4000000',
+  swap: {
+    slippage: '0.4',
+    amount: '40',
+    route: [
+      {
+        token_out_denom: aUSDC,
+        pool_id: '1',
+      },
+    ],
+  },
+  unzapAmount: '1000000',
+  withdrawAmount: '12',
+  zap: {
+    coinsIn: [
+      {
+        denom: aUSDC,
+        amount: '1',
+      },
+      { denom: uosmo, amount: '3' },
+    ],
+    denomOut: usdcOsmo,
+  },
+}
+
 export const osmosisTestnetConfig: DeploymentConfig = {
-  creditLineCoins: [
-    { denom: uosmo, creditLine: defaultCreditLine },
-    { denom: aUSDC, creditLine: defaultCreditLine },
-    { denom: ausdcOsmoPool, creditLine: defaultCreditLine },
-  ],
+  mainnet: false,
+  deployerMnemonic: 'TO BE INSERTED AT TIME OF DEPLOYMENT',
+  marsDenom: mars,
+  atomDenom: atom,
+  safetyFundAddr: safetyFundAddr,
+  protocolAdminAddr: protocolAdminAddr,
+  feeCollectorAddr: feeCollectorAddr,
   chain: {
     baseDenom: uosmo,
     defaultGasPrice: 0.1,
@@ -42,70 +249,45 @@ export const osmosisTestnetConfig: DeploymentConfig = {
     prefix: 'osmo',
     rpcEndpoint: 'https://rpc.osmotest5.osmosis.zone',
   },
-  deployerMnemonic:
-    'rely wonder join knock during sudden slow plate segment state agree also arrest mandate grief ordinary lonely lawsuit hurt super banana rule velvet cart',
-  maxUnlockingPositions: '10',
+  oracle: {
+    name: 'osmosis',
+    baseDenom: 'uosmo',
+  },
+  rewardsCollector: {
+    name: 'osmosis',
+    timeoutSeconds: 600,
+    channelId: 'channel-2083',
+    safetyFundFeeShare: '0.5',
+    feeCollectorDenom: mars,
+    safetyFundDenom: aUSDC,
+    slippageTolerance: '0.01',
+  },
+  incentives: {
+    epochDuration: 604800, // 1 week
+    maxWhitelistedIncentiveDenoms: 10,
+  },
+  swapper: {
+    name: 'osmosis',
+    routes: [
+      { denom_in: atom, denom_out: uosmo, route: [{ pool_id: 12, token_out_denom: uosmo }] },
+      { denom_in: uosmo, denom_out: atom, route: [{ pool_id: 12, token_out_denom: atom }] },
+      { denom_in: aUSDC, denom_out: uosmo, route: [{ pool_id: 5, token_out_denom: uosmo }] },
+      { denom_in: uosmo, denom_out: aUSDC, route: [{ pool_id: 5, token_out_denom: aUSDC }] },
+    ],
+  },
+  targetHealthFactor: '1.05',
+  creditLineCoins: [
+    { denom: uosmo, creditLine: defaultCreditLine },
+    { denom: aUSDC, creditLine: defaultCreditLine },
+    { denom: usdcOsmo, creditLine: defaultCreditLine },
+  ],
+  maxValueForBurn: '10000',
+  maxUnlockingPositions: '1',
   maxSlippage: '0.2',
-  maxValueForBurn: '1000000',
-  // Latest from: https://github.com/mars-protocol/outposts/blob/master/scripts/deploy/addresses/osmo-test-5.json
-  addressProvider: { addr: 'osmo1wlm6dc0vnncu2v5z26rv97plmlkmalm84uwqatrlftc4gmp8ahgqs6r4py' },
-  redBank: { addr: 'osmo1hs4sm0fah9rk4mz8e56v4n76g0q9fffdkkjm3f8tjagkdx78pqcq75pk0a' },
-  incentives: { addr: 'osmo1nu0k6g294jela67vyth6nwr3l42gutq2m07pg9927f7v7tuv0d4sre9fr7' },
-  oracle: { addr: 'osmo1dxu93scjdnx42txdp9d4hm3snffvnzmkp4jpc9sml8xlu3ncgamsl2lx58' },
-  swapper: { addr: 'osmo1ee9cq8dcknmw43znznx6vuupx5ku0tt505agccgaz5gn48mhe45s3kwwfm' },
-  params: { addr: 'osmo1h334tvddn82m4apm08rm9k6kt32ws7vy0c4n30ngrvu6h6yxh8eq9l9jfh' },
-  rewardsCollector: { addr: 'osmo1h334tvddn82m4apm08rm9k6kt32ws7vy0c4n30ngrvu6h6yxh8eq9l9jfh' },
-  // Latest from: https://api.apollo.farm/api/graph?query=query+MyQuery+%7B%0A++vaults%28network%3A+osmo_test_5%29+%7B%0A++++label%0A++++contract_address%0A++%7D%0A%7D
-  vaults: [aUSDC_OSMO_Config(ausdcOsmoVault), ATOM_OSMO_Config(atomOsmoVault)],
   zapperContractName: 'mars_zapper_osmosis',
   runTests: true,
-  testActions: {
-    vault: {
-      depositAmount: '1000000',
-      withdrawAmount: '1000000',
-      mock: {
-        config: {
-          deposit_cap: { denom: aUSDC, amount: '100000000' }, // 100 usdc
-          liquidation_threshold: '0.585',
-          max_loan_to_value: '0.569',
-          whitelisted: true,
-        },
-        vaultTokenDenom: uosmo,
-        type: VaultType.LOCKED,
-        lockup: { time: 900 }, // 15 mins
-        baseToken: ausdcOsmoPool,
-      },
-    },
-    outpostsDeployerMnemonic:
-      'elevator august inherit simple buddy giggle zone despair marine rich swim danger blur people hundred faint ladder wet toe strong blade utility trial process',
-    borrowAmount: '10',
-    repayAmount: '11',
-    depositAmount: '100',
-    lendAmount: '10',
-    reclaimAmount: '5',
-    secondaryDenom: aUSDC,
-    startingAmountForTestUser: '4000000',
-    swap: {
-      slippage: '0.4',
-      amount: '40',
-      route: [
-        {
-          token_out_denom: aUSDC,
-          pool_id: '1',
-        },
-      ],
-    },
-    unzapAmount: '1000000',
-    withdrawAmount: '12',
-    zap: {
-      coinsIn: [
-        {
-          denom: aUSDC,
-          amount: '1',
-        },
-        { denom: uosmo, amount: '3' },
-      ],
-      denomOut: ausdcOsmoPool,
-    },
-  },
+  testActions: testActions,
+  assets: [osmoAsset, atomAsset, USDCAsset],
+  vaults: [usdcOsmoVault, atomOsmoVault],
+  oracleConfigs: [osmoOracle, atomOracle, USDCOracle, atomOsmoOracle, usdcOsmoOracle],
 }
