--- scripts/deploy/osmosis/testnet-deployer.ts
@@ -1,9 +1,0 @@
-import { taskRunner } from '../base/index.rover'
-import { osmosisTestnetConfig } from './testnet-config'
-
-void (async function () {
-  await taskRunner({
-    config: osmosisTestnetConfig,
-    label: 'testnet-deployer-owner',
-  })
-})()
