+++ scripts/deploy/osmosis/testnet-index.ts
@@ -1,5 +1,5 @@
 import { taskRunner } from '../base'
-import { osmosisTestnetConfig } from './testnetConfig'
+import { osmosisTestnetConfig } from './testnet-config'
 
 void (async function () {
   await taskRunner(osmosisTestnetConfig)
